[libsocialcache] Initial work on Facebook picture cache
This commit provides the first work on libsocialcache.
libsocialcache is a library that is used to save and
load data that comes from social networks, and are
cached locally.

libsocialcache includes a library that provides classes
that wraps sqlite databases. While org.nemomobile.socialcache
QML import provides higher level models that can be used
to access the cache.

Right now, only pictures cache from Facebook is
implemented, and Qt4 implementation is not working.