More work on models
This commit includes more work on models, especially on
Twitter posts model. It also includes fixes, like in
picture database, where images are sorted DESC only
in "All" albums.

This commit also includes an optional SyncHelper component
(that can be disabled for desktop builds by passing nodeps
to qmake). This component can be used to track if a social
sync is going on, and adapt the display if needed.