[facebook] Summation of pictures in albums to calculate the pictures per user. Fixes MER#1193
All the pictures in Facebook belong to a certain album. A picture
doesn't belong to two albums.

Hence, we can calculate the total amount of pictures for a user just
by the summation of the claimed pictures contained per user's albums.

This is faster than counting the total amount of pictures per user and
it is more consistent in case we don't have all the pictures for the
albums in the DB or the claimed pictures per album is different than
the amount of pictures returned by Facebook.

The latter would be a bug in Facebook but it won't be so evident for
us, this way.